CloudFunctions_alert/requirements.txt
# Function dependencies, for example:
# package>=version
redis
pyyaml
requests
